Add characteristic, link and arrangement type ids into select lists

+++ LibiadaWeb/Models/Repositories/Catalogs/FullCharacteristicRepository.cs
@@ -168,14 +168,15 @@
             {
                 List<SelectListItem> linkSelectListItems = characteristicsLinks
                     .Where(cl => cl.FullCharacteristic == characteristic && links.Contains(cl.Link))
-                    .Select(cl => new SelectListItem { Value = cl.Link.ToString(), Text = cl.Link.GetDisplayValue() })
+                    .Select(cl => new SelectListItem { Value = ((byte)cl.Link).ToString(), Text = cl.Link.GetDisplayValue() })
                     .Distinct(new SelectListItemComparer())
                     .ToList();
                 List<SelectListItem> arrangementTypeSelectListItems = characteristicsLinks
                     .Where(cl => cl.FullCharacteristic == characteristic && arrangementTypes.Contains(cl.ArrangementType))
-                    .Select(cl => new SelectListItem { Value = cl.ArrangementType.ToString(), Text = cl.ArrangementType.GetDisplayValue() })
+                    .Select(cl => new SelectListItem { Value = ((byte)cl.ArrangementType).ToString(), Text = cl.ArrangementType.GetDisplayValue() })
                     .Distinct(new SelectListItemComparer())
                     .ToList();
+
                 result.Add(new CharacteristicSelectListItem((byte)characteristic, characteristic.GetDisplayValue(), linkSelectListItems, arrangementTypeSelectListItems));
             }
